Delete machine_type.h/cpp
It didn't make sense to be burning a whole TU for one function.

Also unwrap "DllInfo" and "LibInfo" structs that were not providing value.
