+++ src/vcpkg/base/cofffilereader.cpp
@@ -10,7 +10,7 @@ using namespace std;
 
 // See https://docs.microsoft.com/en-us/windows/win32/debug/pe-format
 
-namespace vcpkg::CoffFileReader
+namespace vcpkg
 {
     static void read_and_verify_pe_signature(const ReadFilePointer& fs)
     {
@@ -43,12 +43,12 @@ namespace vcpkg::CoffFileReader
 
     static_assert(sizeof(CoffFileHeader) == 20, "The CoffFileHeader struct must match its on-disk representation");
 
-    DllInfo read_dll(const ReadFilePointer& fs)
+    MachineType read_dll_machine_type(const ReadFilePointer& fs)
     {
         read_and_verify_pe_signature(fs);
         CoffFileHeader header;
         Checks::check_exit(VCPKG_LINE_INFO, fs.read(&header, sizeof(header), 1) == 1);
-        return {to_machine_type(header.machine)};
+        return to_machine_type(header.machine);
     }
 
     struct ArchiveMemberHeader
@@ -193,11 +193,45 @@ namespace vcpkg::CoffFileReader
         return machine_types;
     }
 
-    LibInfo read_lib(const ReadFilePointer& fs)
+    std::vector<MachineType> read_lib_machine_types(const ReadFilePointer& fs)
     {
         read_and_verify_archive_file_signature(fs);
         read_and_skip_first_linker_member(fs);
         const auto offsets = read_second_linker_member_offsets(fs);
-        return {read_machine_types_from_archive_members(fs, offsets)};
+        return read_machine_types_from_archive_members(fs, offsets);
+    }
+
+    MachineType to_machine_type(const uint16_t value)
+    {
+        const MachineType t = static_cast<MachineType>(value);
+        switch (t)
+        {
+            case MachineType::UNKNOWN:
+            case MachineType::AM33:
+            case MachineType::AMD64:
+            case MachineType::ARM:
+            case MachineType::ARM64:
+            case MachineType::ARMNT:
+            case MachineType::EBC:
+            case MachineType::I386:
+            case MachineType::IA64:
+            case MachineType::M32R:
+            case MachineType::MIPS16:
+            case MachineType::MIPSFPU:
+            case MachineType::MIPSFPU16:
+            case MachineType::POWERPC:
+            case MachineType::POWERPCFP:
+            case MachineType::R4000:
+            case MachineType::RISCV32:
+            case MachineType::RISCV64:
+            case MachineType::RISCV128:
+            case MachineType::SH3:
+            case MachineType::SH3DSP:
+            case MachineType::SH4:
+            case MachineType::SH5:
+            case MachineType::THUMB:
+            case MachineType::WCEMIPSV2: return t;
+            default: Checks::exit_maybe_upgrade(VCPKG_LINE_INFO, "Unknown machine type code 0x%hx", value);
+        }
     }
 }
